requirements.txt
aiogram==2.25.1
openpyxl==3.1.2
pandas==2.0.1
python-dotenv==1.0.1
pytz==2023.3
XlsxWriter==3.2.0
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0